+++ tests/test_backend_slurm.py
@@ -1,5 +1,6 @@
 import shlex
 import subprocess
+import time
 from concurrent.futures import Executor
 from itertools import product
 from typing import Callable
@@ -14,6 +15,7 @@ from .fixtures_tasks import MockWorkflowTask
 from fractal_server.app.runner._slurm import SlurmConfig
 from fractal_server.app.runner._slurm.executor import FractalSlurmExecutor
 from fractal_server.app.runner.common import JobExecutionError
+from fractal_server.app.runner.common import TaskExecutionError
 from fractal_server.tasks import dummy as dummy_module
 from fractal_server.tasks import dummy_parallel as dummy_parallel_module
 
@@ -105,7 +107,7 @@ def test_unit_sbatch_script_readable(
     assert "This does not look like a batch script" in out.stderr
 
 
-def test_slurm_executor(
+def test_slurm_executor_submit(
     monkey_slurm,
     monkey_slurm_user,
     tmp777_path,
@@ -121,13 +123,105 @@ def test_slurm_executor(
         slurm_user=monkey_slurm_user,
         working_dir=tmp777_path,
         working_dir_user=tmp777_path,
-        slurm_poll_interval=4,
+        slurm_poll_interval=2,
     ) as executor:
         res = executor.submit(lambda: 42)
     assert res.result() == 42
 
 
-def test_slurm_executor_separate_folders(
+def test_slurm_executor_submit_with_exception(
+    monkey_slurm,
+    monkey_slurm_user,
+    tmp777_path,
+    cfut_jobs_finished,
+):
+    """
+    GIVEN a docker slurm cluster and a FractalSlurmExecutor executor
+    WHEN a function is submitted to the executor and raises an exception
+    THEN the executor raises a TaskExecutionError
+    """
+
+    def raise_ValueError():
+        raise ValueError
+
+    with pytest.raises(TaskExecutionError) as e:
+        with FractalSlurmExecutor(
+            slurm_user=monkey_slurm_user,
+            working_dir=tmp777_path,
+            working_dir_user=tmp777_path,
+            slurm_poll_interval=2,
+        ) as executor:
+            fut = executor.submit(raise_ValueError)
+            debug(fut.result())
+    debug(e.value)
+
+
+def test_slurm_executor_map(
+    monkey_slurm,
+    monkey_slurm_user,
+    tmp777_path,
+    cfut_jobs_finished,
+):
+    with FractalSlurmExecutor(
+        slurm_user=monkey_slurm_user,
+        working_dir=tmp777_path,
+        working_dir_user=tmp777_path,
+        slurm_poll_interval=2,
+    ) as executor:
+        result_generator = executor.map(lambda x: 2 * x, range(4))
+        results = list(result_generator)
+        debug(results)
+        assert results == [2 * x for x in range(4)]
+
+
+@pytest.mark.parametrize("early_late", ["early", "late"])
+def test_slurm_executor_map_with_exception(
+    early_late,
+    monkey_slurm,
+    monkey_slurm_user,
+    tmp777_path,
+    cfut_jobs_finished,
+):
+
+    """
+    NOTE: Tasks submitted to FractalSlurmExecutor via fractal-server always
+    return either JobExecutionError or TaskExecutionError, while for functions
+    submitted directly to the executor (and raising arbitrary exceptions like a
+    ValueError) this is not true. Depending on the way the error is raised, in
+    this test, the resulting error could be a JobExecutionError or
+    TaskExecutionError; we accept both of them, here.
+    """
+
+    debug(early_late)
+
+    def _raise(n: int):
+        if n == 1:
+            if early_late == "late":
+                time.sleep(1.5)
+            raise ValueError
+        else:
+            if early_late == "early":
+                time.sleep(1.5)
+            return n
+
+    with FractalSlurmExecutor(
+        slurm_user=monkey_slurm_user,
+        working_dir=tmp777_path,
+        working_dir_user=tmp777_path,
+        slurm_poll_interval=1,
+    ) as executor:
+        try:
+            result_generator = executor.map(_raise, range(4))
+            for result in result_generator:
+                debug(f"While looping over results, I got to {result=}")
+            raise RuntimeError("We should never reach this line")
+        except Exception as e:
+            debug(e)
+            debug(vars(e))
+            assert type(e) in [TaskExecutionError, JobExecutionError]
+
+
+def test_slurm_executor_submit_separate_folders(
     monkey_slurm,
     monkey_slurm_user,
     tmp777_path,
@@ -146,13 +240,13 @@ def test_slurm_executor_separate_folders(
         slurm_user=monkey_slurm_user,
         working_dir=server_working_dir,
         working_dir_user=user_working_dir,
-        slurm_poll_interval=4,
+        slurm_poll_interval=2,
     ) as executor:
         res = executor.submit(lambda: 42)
     assert res.result() == 42
 
 
-def test_slurm_executor_scancel(
+def test_slurm_executor_submit_and_scancel(
     monkey_slurm,
     monkey_slurm_user,
     tmp777_path,
@@ -181,7 +275,7 @@ def test_slurm_executor_scancel(
             working_dir_user=user_working_dir,
             debug=True,
             keep_logs=True,
-            slurm_poll_interval=4,
+            slurm_poll_interval=2,
         ) as executor:
             fut = executor.submit(wait_and_return)
             debug(fut)