+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Add test for custom tasks (\#500).
 * Improve formatting of job logs (\#503).
+* Fix handling of parallel-tasks errors in `FractalSlurmExecutor` (\#497).
 
 # 1.0.7